backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic>=0.25.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
fastapi==0.109.0
ffmpeg-python==0.2.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gotrue==1.3.1
h11==0.14.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==0.17.3
httptools==0.6.4
httpx>=0.25.0
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lxml==6.0.1
MarkupSafe==3.0.2
marshmallow==3.26.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
openai
packaging==23.2
pandas==2.2.0
pillow==11.3.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
postgrest
proglog==0.1.12
propcache==0.3.2
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pypdf==4.0.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
realtime==1.0.6
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.4.0
scipy==1.16.1
sentence-transformers==2.3.1
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
storage3
StrEnum==0.4.15
supabase
supabase_functions
supafunc==0.2.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==12.0
xlsxwriter==3.2.5
yarl==1.20.1
